[EN SD] Use new pedestal-toolbox cors matcher.
This is a breaking change for the way we configure our servers. We want
to allow regexes for matching CORS Origin headers instead of using
String equality.

This PR should be reviewed well because it will be used as a model to
change all http-api services. We will also need to update Krakenstein's
docker-compose.yml to change the default dev ALLOWED_ORIGINS from
`:all`, which no longer will work. Before deploying any of those
services, we should change the Consul keys for allowed-origins for each
of the services to reflect all subdomains of "turbovote.org". The new
value should be `["https://(.+[.])?turbovote[.]org"]`.